Extract api handler to api namespace
