+++ packages/purge-cache-plugin/src/api/purge-api-handler.ts
@@ -4,13 +4,24 @@ import {
   type PayloadHandler,
   type PayloadRequest,
 } from 'payload';
-import { canAccessPurgeCache } from './access.js';
-import {
-  type PurgeCacheErrorResponse,
-  purgeCacheRequestSchema,
-  type PurgeCacheResponse,
-} from './request-schema.js';
-import type { PurgeCachePluginConfig, PurgerResult } from './types.js';
+import { z, type ZodError } from 'zod';
+import { canAccessPurgeCache } from '../access.js';
+import type { PurgeCachePluginConfig, PurgerResult } from '../types.js';
+
+export const purgeCacheRequestSchema = z.object({
+  purge: z.array(z.string()),
+});
+
+export type PurgeCacheRequestData = z.infer<typeof purgeCacheRequestSchema>;
+
+export type PurgeCacheResponse = {
+  results: Record<string, PurgerResult>;
+};
+
+export type PurgeCacheErrorResponse = {
+  error: string;
+  violations?: ZodError['issues'];
+};
 
 const response = (
   req: PayloadRequest,
